Had a "windows session" with Neo and fixed a few problems. Most often these problems where related to some file channel/memory mapping not beeing released properly. File channels are now closed properly and if a file operation fails because of a memory mapped region hasn't been gc:ed yet a System.gc() will be invoked and the file operation will be tried again.
Found one nasty bug regarding FileChannel position. Apparently using a memory mapped 
buffer doesn't modify the underlying FileChannel position on windows the same way it does 
on a unix OS. Fixed this in LogBuffer so file channel position is tracked in the 
implementation.

Also took the opportunity to fix some warnings that latest vanilla eclipse version 
complained about.


git-svn-id: https://svn.neo4j.org/projects/neo/trunk@411 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ advanced/src/java/org/neo4j/impl/transaction/xaframework/XaLogicalLog.java
@@ -217,6 +217,10 @@ public class XaLogicalLog
 		buffer.flip();
 		int identifier = buffer.getInt();
 		Xid xid = xidIdentMap.get( identifier );
+		if ( xid == null )
+		{
+			return false;
+		}
 		if ( xaRm.injectPrepare( xid ) )
 		{
 			// read only we can remove
@@ -253,6 +257,10 @@ public class XaLogicalLog
 		buffer.flip();
 		int identifier = buffer.getInt();
 		Xid xid = xidIdentMap.get( identifier );
+		if ( xid == null )
+		{
+			return false;
+		}
 		xaRm.injectOnePhaseCommit( xid );
 		return true;
 	}
@@ -294,6 +302,10 @@ public class XaLogicalLog
 		buffer.flip();
 		int identifier = buffer.getInt();
 		Xid xid = xidIdentMap.get( identifier );
+		if ( xid == null )
+		{
+			return false;
+		}
 		xaRm.pruneXid( xid );
 		xidIdentMap.remove( identifier );
 		recoveredTxMap.remove( identifier );
@@ -404,11 +416,18 @@ public class XaLogicalLog
 		{
 			log.info( "Active transactions: " + xidIdentMap.size() );
 			log.info( "Closing dirty log: " + fileName );
-			writeBuffer.force();
+			if ( writeBuffer != null )
+			{
+				writeBuffer.force();
+			}
+			writeBuffer = null;
 			fileChannel.close();
 			return;
 		}
-		writeBuffer.force();
+		if ( writeBuffer != null )
+		{
+			writeBuffer.force();
+		}
 		writeBuffer = null;
 		fileChannel.close();
 		File file = new File( fileName );
@@ -417,25 +436,26 @@ public class XaLogicalLog
 			throw new IOException( "Logical log[" + fileName + "] not found" );
 		}
 		// TODO: if store old logs save them here
+		boolean deleted = false;
 		try
 		{
-			file.delete();
-		}
-		catch ( Exception e )
+			deleted = file.delete();
+		} catch ( Exception e ) {}
+		if ( !deleted )
 		{
+			// hack for WINBLOWS
 			try
 			{
-				Thread.sleep( 500 );
-				// try again
-				file.delete();
-			}
-			catch ( Throwable t )
-			{
-				log.warning( "Unable to delete clean logical log[" +  
-					fileName + "]" + e );
+				System.gc();
+				deleted = file.delete();
+			} catch ( Exception ee ) {}
+		}
+		if ( !deleted )
+		{
+			log.warning( "Unable to delete clean logical log[" +  
+					fileName + "]" );
 				log.info( "Since you're not using a real operating system " + 
 					"you will be punished for it at next startup" );
-			}
 		}
 	}
